Extract presel xml parser in to separate class
